removed redundant cast
git-svn-id: svn+ssh://developer.mitre.org/svnroot/giscore/trunk@1221 5c99b798-f2dd-11dd-9a12-8f84353f4b73

+++ giscore/src/org/mitre/giscore/impl/BlockInputStream.java
@@ -54,7 +54,7 @@ public class BlockInputStream extends InputStream {
 	 */
 	public int read() throws IOException {
 		if (checkBuffer())
-			return (int) (0xFF & buffer[pos++]);
+			return (0xFF & buffer[pos++]);
 		else
 			return -1;
 	}
